Add DI framework as dependency
Use the DependencyInjector framework to inject containers in the routers
which will be responsible for instantiating the whole flow of the app.

The Graph's container should register and resolve all dependencies of 
the app. In order to do it the container must be passed down to every 
Router.
No other classes besides Routers should have access to it. Routers are 
the only responsible for instantiating new screens and they are 
in charge of injecting all necessary dependencies into them.